move global configuration
