+++ internal/k8s/global_configuration.go
@@ -1,0 +1,125 @@
+/*
+Copyright 2015 The Kubernetes Authors All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package k8s
+
+import (
+	"fmt"
+	"reflect"
+
+	"github.com/golang/glog"
+	conf_v1 "github.com/nginxinc/kubernetes-ingress/pkg/apis/configuration/v1"
+	api_v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/fields"
+	"k8s.io/client-go/tools/cache"
+)
+
+func createGlobalConfigurationHandlers(lbc *LoadBalancerController) cache.ResourceEventHandlerFuncs {
+	return cache.ResourceEventHandlerFuncs{
+		AddFunc: func(obj interface{}) {
+			gc := obj.(*conf_v1.GlobalConfiguration)
+			glog.V(3).Infof("Adding GlobalConfiguration: %v", gc.Name)
+			lbc.AddSyncQueue(gc)
+		},
+		DeleteFunc: func(obj interface{}) {
+			gc, isGc := obj.(*conf_v1.GlobalConfiguration)
+			if !isGc {
+				deletedState, ok := obj.(cache.DeletedFinalStateUnknown)
+				if !ok {
+					glog.V(3).Infof("Error received unexpected object: %v", obj)
+					return
+				}
+				gc, ok = deletedState.Obj.(*conf_v1.GlobalConfiguration)
+				if !ok {
+					glog.V(3).Infof("Error DeletedFinalStateUnknown contained non-GlobalConfiguration object: %v", deletedState.Obj)
+					return
+				}
+			}
+			glog.V(3).Infof("Removing GlobalConfiguration: %v", gc.Name)
+			lbc.AddSyncQueue(gc)
+		},
+		UpdateFunc: func(old, cur interface{}) {
+			curGc := cur.(*conf_v1.GlobalConfiguration)
+			if !reflect.DeepEqual(old, cur) {
+				glog.V(3).Infof("GlobalConfiguration %v changed, syncing", curGc.Name)
+				lbc.AddSyncQueue(curGc)
+			}
+		},
+	}
+}
+
+func (lbc *LoadBalancerController) addGlobalConfigurationHandler(handlers cache.ResourceEventHandlerFuncs, namespace string, name string) {
+	lbc.globalConfigurationLister, lbc.globalConfigurationController = cache.NewInformer(
+		cache.NewListWatchFromClient(
+			lbc.confClient.K8sV1().RESTClient(),
+			"globalconfigurations",
+			namespace,
+			fields.Set{"metadata.name": name}.AsSelector()),
+		&conf_v1.GlobalConfiguration{},
+		lbc.resync,
+		handlers,
+	)
+	lbc.cacheSyncs = append(lbc.cacheSyncs, lbc.globalConfigurationController.HasSynced)
+}
+
+func (lbc *LoadBalancerController) syncGlobalConfiguration(task task) {
+	key := task.Key
+	obj, gcExists, err := lbc.globalConfigurationLister.GetByKey(key)
+	if err != nil {
+		lbc.syncQueue.Requeue(task, err)
+		return
+	}
+
+	var changes []ResourceChange
+	var problems []ConfigurationProblem
+	var validationErr error
+
+	if !gcExists {
+		glog.V(2).Infof("Deleting GlobalConfiguration: %v\n", key)
+
+		changes, problems = lbc.configuration.DeleteGlobalConfiguration()
+	} else {
+		glog.V(2).Infof("Adding or Updating GlobalConfiguration: %v\n", key)
+
+		gc := obj.(*conf_v1.GlobalConfiguration)
+		changes, problems, validationErr = lbc.configuration.AddOrUpdateGlobalConfiguration(gc)
+	}
+
+	updateErr := lbc.processChangesFromGlobalConfiguration(changes)
+
+	if gcExists {
+		eventTitle := "Updated"
+		eventType := api_v1.EventTypeNormal
+		eventMessage := fmt.Sprintf("GlobalConfiguration %s was added or updated", key)
+
+		if validationErr != nil {
+			eventTitle = "AddedOrUpdatedWithError"
+			eventType = api_v1.EventTypeWarning
+			eventMessage = fmt.Sprintf("GlobalConfiguration %s is updated with errors: %v", key, validationErr)
+		}
+
+		if updateErr != nil {
+			eventTitle += "WithError"
+			eventType = api_v1.EventTypeWarning
+			eventMessage = fmt.Sprintf("%s; with reload error: %v", eventMessage, updateErr)
+		}
+
+		gc := obj.(*conf_v1.GlobalConfiguration)
+		lbc.recorder.Eventf(gc, eventType, eventTitle, eventMessage)
+	}
+
+	lbc.processProblems(problems)
+}
