+++ internal/k8s/handlers.go
@@ -171,40 +171,6 @@ func createVirtualServerRouteHandlers(lbc *LoadBalancerController) cache.Resourc
 	}
 }
 
-func createGlobalConfigurationHandlers(lbc *LoadBalancerController) cache.ResourceEventHandlerFuncs {
-	return cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			gc := obj.(*conf_v1.GlobalConfiguration)
-			glog.V(3).Infof("Adding GlobalConfiguration: %v", gc.Name)
-			lbc.AddSyncQueue(gc)
-		},
-		DeleteFunc: func(obj interface{}) {
-			gc, isGc := obj.(*conf_v1.GlobalConfiguration)
-			if !isGc {
-				deletedState, ok := obj.(cache.DeletedFinalStateUnknown)
-				if !ok {
-					glog.V(3).Infof("Error received unexpected object: %v", obj)
-					return
-				}
-				gc, ok = deletedState.Obj.(*conf_v1.GlobalConfiguration)
-				if !ok {
-					glog.V(3).Infof("Error DeletedFinalStateUnknown contained non-GlobalConfiguration object: %v", deletedState.Obj)
-					return
-				}
-			}
-			glog.V(3).Infof("Removing GlobalConfiguration: %v", gc.Name)
-			lbc.AddSyncQueue(gc)
-		},
-		UpdateFunc: func(old, cur interface{}) {
-			curGc := cur.(*conf_v1.GlobalConfiguration)
-			if !reflect.DeepEqual(old, cur) {
-				glog.V(3).Infof("GlobalConfiguration %v changed, syncing", curGc.Name)
-				lbc.AddSyncQueue(curGc)
-			}
-		},
-	}
-}
-
 func createPolicyHandlers(lbc *LoadBalancerController) cache.ResourceEventHandlerFuncs {
 	return cache.ResourceEventHandlerFuncs{
 		AddFunc: func(obj interface{}) {
